+++ api/pmpanel/model.go
@@ -9,8 +9,8 @@ type NodeInfoResponse struct {
 	Method          string  `json:"method"`
 	TrafficRate     float64 `json:"trafficRate"`
 	RawServerString string  `json:"outServer"`
-	Port            int     `json:"outPort"`
-	AlterId         int     `json:"alterId"`
+	Port            uint32  `json:"outPort"`
+	AlterId         uint16  `json:"alterId"`
 	Network         string  `json:"network"`
 	Security        string  `json:"security"`
 	Host            string  `json:"host"`
